+++ HX-ANiMe-Java/src/main/java/com/hx/pojo/DO/forcegraph/NodeDO.java
@@ -20,16 +20,16 @@ public class NodeDO {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer nodeId;
+    private Long nodeId;
 
     @Column(nullable = false)
-    private Integer userId;
+    private Long userId;
 
     @Column(nullable = false)
-    private Integer userTableId;
+    private Long userTableId;
 
     @Column(nullable = false)
-    private Integer legendId;
+    private Long legendId;
 
     @Column(nullable = false, length = 255)
     private String name;
